Plot a fitted cpgam (superseded)

Plot a fitted cpgam (superseded)

Usage

plot_gene_co(
  cpo,
  gene_id = NULL,
  target_id = NULL,
  cp_type = c("cp_1se", "cp_min"),
  shape_type = "shape1",
  bs = "auto",
  cp_fix = -1,
  facet = F,
  sp = NULL,
  show_fit = T,
  show_data = T,
  show_fit_ci = T,
  show_data_ci = T,
  ci_prob = "se",
  remove_null = F,
  null_threshold = 0.05,
  null_threshold_adj = T,
  k_mult = 1.2,
  gene_level_plot = F,
  return_fits_only = F,
  family = "nb",
  common_y_scale = T,
  scaled = F
)

Arguments

cpo

a cpam object

gene_id

character; gene_id

target_id

character; target_id

cp_type

character; if changepoints have been estimated using estimate_changepoint(), which selection rule should be used. See estimate_changepoint() for details.

shape_type

character; if shapes have been estimated using select_shape(), which set of candidate shapes should be used. See select_shape() for details.

bs

character; set the basis (i.e. shape)

cp_fix

numerical; set the changepoint

facet

logical; for genes with multiple transcripts, should the transcripts be plotted in separate facets

sp

numerical; set the smooth parameter

show_fit

logical; show the fitted trend

show_data

logical; show (possibly normalized and scaled) data points

show_fit_ci

logical; show credible interval for the fitted trend

show_data_ci

logical; show bootstrapped quantile for data points

ci_prob

if numerical, sets the probability for the simulation-based estimated of credible interval, if equal to "se", the interval is set to the approximate standard error (see mgcv::predict.gam())

remove_null

logical; only plot differentially expressed transcripts

null_threshold

numeric; P value threshold for filtering out NULL transcripts

null_threshold_adj

logical; use adjusted (default) or non-adjusted p-values for filtering targets

k_mult

numerical; multiplier for the number of knots in the spline

gene_level_plot

logical; plot gene-level data and fitted trend

return_fits_only

logical; return the model fits. Does not plot the function

family

character; negative binomial ("nb", default) or Gaussian ("gaussian")

common_y_scale

logical; for faceted plots of multiple transcripts, should the scale of the y-axis be common or free.

scaled

logical; scaled data by overdispersions (for bootstrapped data only)

Please use plot_cpam() instead. This function will be deprecated in future versions.

Value

a ggplot object

Examples

1+1
#> [1] 2