Create a results table from a cpam object

Create a results table from a cpam object

Usage

results(
  cpo,
  p_threshold = 0.05,
  p_type = c("p_gam", "p_mvn"),
  min_lfc = 0,
  min_count = 0,
  aggregate_to_gene = cpo$aggregate_to_gene,
  add_lfc = T,
  add_counts = T,
  cp_type = c("cp_1se", "cp_min"),
  shape_type = c("shape1", "shape2"),
  summarise_to_gene = F,
  remove_null_targets = T
)

Arguments

cpo

a cpam object

p_threshold

numerical; threshold for adjusted p-values; default is 0.05

p_type

character; choose the type of p-value. Options are "p_gam" (default) or "p_mvn" (see compute_p_values() for details).

min_lfc

numerical; maximum absolute log (base 2) fold change must exceed this minimum value; default is 0

min_count

numerical; maximum of the modelled counts evaluated at the set of observed time points must exceed this minimum value for

aggregate_to_gene

logical; filter by gene-aggregated p-values

add_lfc

logical; add log (base 2) fold changes for each time point

add_counts

logical; add modelled counts for each time point

cp_type

character; model-selection rule used to select the changepoint

shape_type

character; "shape1" to include unconstrained or otherwise "shape2"

summarise_to_gene

logical; return gene-level results only

remove_null_targets

logical; remove targets with null shapes (default is T). If F, targets with null shapes will be included if the aggregated p-value for the corresponding gene passes the specified filtering thresholds.

Value

a tibble

Details

This function is usually called after compute_p_values(), estimate_changepoint, and select_shape have been run. The function has several useful filters such as adjusted p-value thresholds, minimum log-fold changes, and minimum counts.

Examples